This code is older than what is in the CVS repository.  From Mark Rivers's email today:
I have compared the code in ASD CVS repository (/extensions/src/idl) with the BCDA repository
(https://subversion.xray.aps.anl.gov/bcdaext/scanSee/trunk)

There are differences in the IDL .pro files.

It looks to me that the CVS code is more recent.  For example in DC.pro the code in CVS claims to be R3.4.4.2, while in SVN it is R3.3.
...